Title: 세 수의 합과 평균 구하기, Time: 133ms, Memory: 10MB, Status: Passed - Codetree

+++ 260305/세 수의 합과 평균 구하기/sum-and-mean-of-three-numbers.java
@@ -7,7 +7,7 @@ public class Main {
         int a=sc.nextInt();
         int b=sc.nextInt();
         int c =sc.nextInt();
-        double d = (double)(a+b+c)/3;
+        double d = (int)(a+b+c)/3;
 
         System.out.printf("%d\n%.0f",a+b+c,d);
